requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
apscheduler
email-validator
pytz
